Deleted Contributors_Only/Source and moved files
