+++ CMakeLists.txt
@@ -23,12 +23,17 @@ set(SRC
 "src/Generators/MT/MT.cpp"
 "src/Generators/Naor-Reingold/naor_reingold.cpp"
 "src/Generators/XORShift/XORShift.cpp"
+"src/Distributions/Discrete/Bernoulli/Bernoulli.cpp"
 "src/Distributions/Discrete/Binomial/Binomial.cpp"
 "src/Distributions/Discrete/Gibbs/Gibbs.cpp"
 "src/Distributions/Discrete/Hypergeometric/Hypergeometric.cpp"
 "src/Distributions/Discrete/Negative-Hypergeometric/NegHypergeometric.cpp"
+"src/Distributions/Discrete/Poisson/Poisson.cpp"
 "src/Distributions/Continuous/Cauchy/Cauchy.cpp"
-"src/Distributions/Continuous/Exponential/Exponential.cpp")
+"src/Distributions/Continuous/Exponential/Exponential.cpp"
+"src/Distributions/Continuous/Gaussian/Gaussian.cpp"
+"src/Distributions/Continuous/Maxwell/Maxwell.cpp"
+"src/Distributions/Continuous/Weibull/Weibull.cpp")
 
 # Compile to objects
