--- Contributors_Only/Source/CMakeLists.txt
@@ -1,6 +1,0 @@
-cmake_minimum_required(VERSION 3.19)
-
-project(DiceForge VERSION 1.0
-                  DESCRIPTION "A C++ library with the same functionality as the random library in python."
-                  LANGUAGES CXX)
-
