--- Contributors_Only/Source/diceforge.h
@@ -1,7 +1,0 @@
-#ifndef DICEFORGE
-#define DICEFORGE
-
-#include "generators.h"
-#include "distributions.h"
-
-#endif
