--- Contributors_Only/Source/distributions.h
@@ -1,12 +1,0 @@
-#ifndef DF_DISTRIBUTIONS_H
-#define DF_DISTRIBUTIONS_H
-
-#include "Distributions/distribution.h"
-#include "Distributions/Continuous/Cauchy.h"
-#include "Distributions/Continuous/Exponential.h"
-#include "Distributions/Discrete/hypergeometric.h"
-#include "Distributions/Continuous/Weibull.h"
-#include "Distributions/Discrete/NegHypergeometric.h"
-#include "Distributions/Discrete/Gibbs.h"
-
-#endif
