--- Contributors_Only/Testing/bin.dat
@@ -1,5 +1,0 @@
-0	368
-1	380
-2	392
-3	376
-4	383
