+++ src/Distributions/Continuous/Gaussian/Gaussian.cpp
@@ -1,4 +1,5 @@
 #include "Gaussian.h"
+#include <vector>
 
 namespace DiceForge
 {            
@@ -46,7 +47,7 @@ namespace DiceForge
 
     real_t Gaussian::cdf(real_t x) const 
     {   
-        std::vector<int> factorial = {1, 1, 2, 6, 24, 120, 720, 5040, 40320, 362880, 3628800};
+        static const std::vector<int> factorial = {1, 1, 2, 6, 24, 120, 720, 5040, 40320, 362880, 3628800};
         real_t sum = 0; 
         for(int i = 0; i <= 10; i++)
         {
