+++ src/Distributions/Discrete/Bernoulli/Bernoulli.h
@@ -1,8 +1,7 @@
 #ifndef DF_BERNOULLI_H
 #define DF_BERNOULLI_H
 
-#include "../distribution.h"
-
+#include "distribution.h"
 
 namespace DiceForge {
     /// @brief DiceForge::Bernoulli - A Discrete Probability Distribution (Bernoulli) 
@@ -10,11 +9,8 @@ namespace DiceForge {
         private:
             float p;
         public:
-            
-          
             Bernoulli(real_t p);
-
-            int_t next(real_t r) override final;
+            int_t next(real_t r);
             real_t variance() const override final;
             real_t expectation() const override final;
             int_t minValue() const override final;
